Updated to latest UHD (5/9/2011) and fixed residual tuning error.
This commit updates to the most current revision of UHD (current
as of 5/9/2011).  It also incorporates an update to fix a residual
tuning error bug that was discovered with the BURX daughterboard.
This error occurs when tuning to RF frequencies above 2.2 GHz, and
could potentially cause an error of a few 100s of Hz in the final
LO frequency.  This issue is resolved by only tuning the BURX
dboard in steps of 5 MHz, and using the DDCs in the FPGA of the
USRP perform the fine step tuning required to hit the exact LO
frequency requested by the user.