Fix BURX driver to include a sleep after set_bandwidth().
The BURX card needs a minimum of 100 mS after a call to the
set_bandwidth() function in order to guarantee that all
subsequent cmds will be handled successfully by the card.  This
wasn't an issue with non-UHD versions of the driver because the
set_bandwidth() function was always called at the end of the
BURX object constructor, and only once (since there was no
consistent interface exposed for changing this).